# Conflicts:
#	app/src/main/java/com/garnegsoft/hubs/data/HabrApi.kt
#	app/src/main/java/com/garnegsoft/hubs/data/me/MeController.kt

+++ app/src/main/java/com/garnegsoft/hubs/data/me/MeDataUpdateWorker.kt
@@ -30,14 +30,14 @@ class MeDataUpdateWorker(
 			
 			
 			val me = MeController.getMe()
-			if (me == null) {
+			if (me.isSuccess && me.getOrNull() == null) {
 				HubsDataStore.Auth.edit(applicationContext, HubsDataStore.Auth.Alias, "")
 				HubsDataStore.Auth.edit(applicationContext, HubsDataStore.Auth.AvatarFileName, "")
 				HubsDataStore.Auth.edit(applicationContext, HubsDataStore.Auth.LastAvatarUrlDownloaded, "")
 				val avatarFile = File(applicationContext.filesDir, avatarFileName)
 				avatarFile.delete()
 			}
-			me?.let { me ->
+			me.getOrNull()?.let { me ->
 				HubsDataStore.Auth.edit(applicationContext, HubsDataStore.Auth.Alias, me.alias)
 				val newAvatarUrl = me.avatarUrl ?: placeholderAvatarUrl(me.alias)
 				if (lastAvatarUrl != newAvatarUrl || avatarFileName == "" || !File(applicationContext.filesDir, avatarFileName).exists()) {
